+++ src/commonMain/kotlin/io/github/koalaplot/core/xygraph/LinearAxisModel.kt
@@ -7,9 +7,16 @@ internal val TickRatios = listOf(0.1f, 0.2f, 0.5f, 1f, 2f)
 internal const val ZoomRangeLimitDefault = 0.2
 internal const val MinimumMajorTickIncrementDefault = 0.1f
 
-public interface ILinearAxisModel<T> : AxisModel<T> where T : Comparable<T>, T : Number {
+public interface LinearAxisModel<T> : AxisModel<T> where T : Comparable<T>, T : Number {
     public val range: ClosedRange<T>
     public val minimumMajorTickSpacing: Dp
+
+    /**
+     * Sets the currently viewable range on the axis. This is a more direct alternative to using [pan] and [zoom]
+     * to set the range of currently viewable values on an axis. The start and end must be within the allowed
+     * [range] of the axis. If they exceed either limit they will be clamped to the range.
+     */
+    public fun setViewRange(newRange: ClosedRange<T>)
 }
 
 internal fun <X, Y> List<Point<X, Y>>.toXList(): List<X> = object : AbstractList<X>() {